Make AvoidDuplicateElementInitialization language agnostic.
